+++ server/tests-py/test_endpoints.py
@@ -58,3 +58,20 @@ class TestCustomEndpoints:
     def test_endpoint_conflicting(self, hge_ctx, transport):
         check_query_f(hge_ctx, self.dir() + '/endpoint_conflicting.yaml', transport)
 
+    def test_endpoint_duplicate_param(self, hge_ctx, transport):
+        check_query_f(hge_ctx, self.dir() + '/endpoint_duplicate_param.yaml', transport)
+
+    def test_endpoint_empty_path(self, hge_ctx, transport):
+        check_query_f(hge_ctx, self.dir() + '/endpoint_empty_path.yaml', transport)
+        
+    def test_endpoint_trailing_slash(self, hge_ctx, transport):
+        check_query_f(hge_ctx, self.dir() + '/endpoint_trailing_slash.yaml', transport)
+        
+    def test_endpoint_empty_path_segment(self, hge_ctx, transport):
+        check_query_f(hge_ctx, self.dir() + '/endpoint_empty_path_segment.yaml', transport)
+
+    def test_endpoint_empty_path_param(self, hge_ctx, transport):
+        check_query_f(hge_ctx, self.dir() + '/endpoint_empty_path_param.yaml', transport)
+
+    def test_endpoint_subscription(self, hge_ctx, transport):
+        check_query_f(hge_ctx, self.dir() + '/endpoint_subscription.yaml', transport)